refactor: Product를 jpa 엔티티로 변경
