+++ src/main/java/kitchenpos/dao/ProductDao.java
@@ -1,11 +1,12 @@
 package kitchenpos.dao;
 
-import kitchenpos.domain.Product;
-
 import java.util.List;
 import java.util.Optional;
+import kitchenpos.domain.Product.Product;
+import org.springframework.data.repository.Repository;
 
-public interface ProductDao {
+public interface ProductDao extends Repository<Product, Long> {
+
     Product save(Product entity);
 
     Optional<Product> findById(Long id);
